[Authz] Moved ApiPrivileges to core server package (#216051)
## Summary

The `ApiPrivileges` utility was introduced in
https://github.com/elastic/kibana/pull/208067 to help enforce naming
convention standards for API privileges, however the utility is located
in a private platform package and not accessible by non-platform
plugins, e.g. solutions.

Moved `ApiPrivileges` to `src/core/packages/security/server`

### Checklist
- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [x] The PR description includes the appropriate Release Notes section,
and the correct `release_note:*` label is applied per the
[guidelines](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)

__Closes: https://github.com/elastic/kibana/issues/215096__

---------

(cherry picked from commit a19056f5c3f63f1d07564316e782a795fdc32236)

# Conflicts:
#	src/platform/plugins/shared/telemetry/tsconfig.json
#	x-pack/solutions/observability/plugins/profiling/tsconfig.json
