+++ src/core/packages/security/server/index.ts
@@ -50,4 +50,5 @@ export type {
 export type { KibanaPrivilegesType, ElasticsearchPrivilegesType } from './src/roles';
 export { isCreateRestAPIKeyParams } from './src/authentication/api_keys';
 export type { CoreFipsService } from './src/fips';
+export { ApiPrivileges, ApiOperation } from './src/api_privileges';
 export { unwindNestedSecurityPrivileges } from './src/authz';
