+++ src/platform/plugins/shared/telemetry/tsconfig.json
@@ -36,6 +36,7 @@
     "@kbn/core-node-server",
     "@kbn/core-user-profile-browser-mocks",
     "@kbn/security-plugin-types-common",
+    "@kbn/core-security-server",
   ],
   "exclude": [
     "target/**/*",
