+++ x-pack/platform/packages/private/security/authorization_core/src/actions/api.ts
@@ -7,7 +7,7 @@
 
 import { isString } from 'lodash';
 
-import { ApiOperation } from '@kbn/security-plugin-types-common';
+import { ApiOperation } from '@kbn/core-security-server';
 import type { ApiActions as ApiActionsType } from '@kbn/security-plugin-types-server';
 
 export class ApiActions implements ApiActionsType {
