+++ x-pack/platform/packages/private/security/authorization_core/src/privileges/privileges.ts
@@ -7,6 +7,7 @@
 
 import { uniq } from 'lodash';
 
+import { ApiOperation } from '@kbn/core-security-server';
 import type {
   FeatureKibanaPrivileges,
   FeatureKibanaPrivilegesReference,
@@ -17,7 +18,6 @@ import {
   isMinimalPrivilegeId,
 } from '@kbn/security-authorization-core-common';
 import type { RawKibanaPrivileges, SecurityLicense } from '@kbn/security-plugin-types-common';
-import { ApiOperation } from '@kbn/security-plugin-types-common';
 
 import { featurePrivilegeBuilderFactory } from './feature_privilege_builder';
 import type { Actions } from '../actions';
