+++ x-pack/platform/packages/private/security/authorization_core/tsconfig.json
@@ -13,5 +13,6 @@
     "@kbn/security-plugin-types-common",
     "@kbn/security-plugin-types-server",
     "@kbn/licensing-plugin",
+    "@kbn/core-security-server",
   ]
 }
