+++ x-pack/platform/packages/private/security/authorization_core_common/tsconfig.json
@@ -6,7 +6,5 @@
   },
   "include": ["**/*.ts", "**/*.tsx"],
   "exclude": ["target/**/*"],
-  "kbn_references": [
-    "@kbn/security-plugin-types-common",
-  ]
+  "kbn_references": []
 }
