+++ x-pack/platform/packages/shared/security/plugin_types_common/index.ts
@@ -47,5 +47,3 @@ export type {
   CategorizedApiKey,
   ApiKeyAggregations,
 } from './src/api_keys/api_key';
-
-export { ApiOperation } from './src/authorization';
