+++ x-pack/platform/packages/shared/security/plugin_types_common/src/authorization/index.ts
@@ -17,4 +17,3 @@ export type {
   RoleRemoteIndexPrivilege,
   RoleRemoteClusterPrivilege,
 } from './role';
-export { ApiOperation } from './api';
