+++ x-pack/platform/plugins/shared/ai_infra/product_doc_base/tsconfig.json
@@ -26,6 +26,6 @@
     "@kbn/licensing-plugin",
     "@kbn/task-manager-plugin",
     "@kbn/inference-common",
-    "@kbn/security-authorization-core-common",
+    "@kbn/core-security-server",
   ]
 }
