+++ x-pack/platform/plugins/shared/observability_ai_assistant/tsconfig.json
@@ -51,9 +51,9 @@
     "@kbn/server-route-repository-utils",
     "@kbn/inference-plugin",
     "@kbn/ai-assistant-icon",
-    "@kbn/security-authorization-core-common",
     "@kbn/core-http-browser",
-    "@kbn/sse-utils"
+    "@kbn/sse-utils",
+    "@kbn/core-security-server"
   ],
   "exclude": ["target/**/*"]
 }
