+++ x-pack/solutions/observability/plugins/profiling/tsconfig.json
@@ -55,7 +55,8 @@
     "@kbn/deeplinks-observability",
     "@kbn/react-kibana-context-render",
     "@kbn/apm-data-access-plugin",
-    "@kbn/security-plugin-types-common"
+    "@kbn/security-plugin-types-common",
+    "@kbn/core-security-server"
     // add references to other TypeScript projects the plugin depends on
 
     // requiredPlugins from ./kibana.json
